Cleaned and Refactored, moved redux and router to App.js
