+++ frontend/src/App.js
@@ -1,12 +1,41 @@
 import React, { Component } from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { createStore, applyMiddleware } from 'redux';
+import { Provider } from 'react-redux';
+import ReduxPromise from 'redux-promise';
+import Reducers from './reducers';
+
 import './App.css';
+// Common Components
+import Header from './components/Header/Header';
+// Views for Authentication
+import Signin from './components/Auth/Signin';
+import Signup from './components/Auth/Signup';
+import forgotpassword from './components/Auth/forgotpassword';
+// General View Pages
+import Landing from './components/Landing/Landing';
+import Pricing from './components/Pricing/Pricing';
+
+const ReduxStore = createStore(Reducers, applyMiddleware(ReduxPromise));
 
 class App extends Component {
+  state = {}
   render() {
     return (
-      <div className="App">
-        Welcome to Housecups.
-      </div>
+      <Provider store={ReduxStore}>
+        <Router>
+          <div className="App">
+            <Header />
+            <Switch>
+              <Route exact path="/" component={Landing} />
+              <Route exact path="/pricing" component={Pricing} />
+              <Route exact path="/signin" component={Signin} />
+              <Route exact path="/signup" component={Signup} />
+              <Route exact path="/forgotPassword" component={forgotpassword} />
+            </Switch>
+          </div>
+        </Router>
+      </Provider>
     );
   }
 }
