+++ frontend/src/components/Header/Header.js
@@ -4,8 +4,6 @@ import './Header.css';
 
 class Header extends Component {
 
-  state = {}
-
   render() {
     return (
       <div className="Header">
