+++ frontend/src/components/Landing/Landing.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import './Landing.css';
 import Section from '../Section/Section';
-import Pricing from '../Pricing/Pricing';
 
 class Landing extends Component {
 
@@ -22,13 +21,9 @@ class Landing extends Component {
               <button>Create your School</button>
             </div>
           </div>
-          <div className="Intro__image">
-          
-          </div>
+          <div className="Intro__image"></div>
 
         </Section>
-
-        <Pricing />
 
       </div>
     );
